# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexWatch Hub API
  description: >
    Custom API routes exposed by the NexWatch hub (a PocketBase application)
    under /api/custom/*, plus the unauthenticated /healthz liveness probe.
    PocketBase's own built-in routes (record CRUD under /api/collections/*,
    auth, realtime, files, etc.) are not documented here — see the
    PocketBase API reference for those.
  version: "1.0.0"
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/custom/reports/preview:
    get:
      operationId: previewReport
      summary: Render the weekly fleet report without sending it
      description: >
        Builds the report for the requested period and returns its
        rendered subject, HTML body, and plain-text body, without
        delivering it through any notification channel. Requires the
        "admin" role.
      tags: [reports]
      parameters:
        - name: period_days
          in: query
          schema: { type: integer, minimum: 1 }
          description: Lookback window in days. Defaults to the report_period_days setting (7).
      responses:
        "200":
          description: The rendered report.
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject: { type: string }
                  html: { type: string }
                  text: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/custom/reports/send:
    post:
      operationId: sendReportNow
      summary: Build and send the weekly fleet report now
      description: >
        Builds the report for the configured report_period_days and
        delivers it through every channel in report_channel_ids —
        skipping (with a per-channel error) any channel that is missing,
        disabled, or not of type "email". Requires the "admin" role.
      tags: [reports]
      responses:
        "200":
          description: Delivery results, one entry per configured channel.
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject: { type: string }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        channel_id: { type: string }
                        channel_name: { type: string }
                        success: { type: boolean }
                        error: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/custom/push/vapid-public-key:
    get:
      operationId: getPushVAPIDPublicKey
      summary: Get the hub's VAPID public key for Web Push subscriptions
      description: >
        Returns the hub's VAPID public key, used by the browser's
        PushManager.subscribe() call. The key pair is generated lazily on
        the first request (and persisted — the private key in the
        superuser-only "hub_secrets" collection, the public key in the
        "vapid_public_key" setting), so every subsequent call returns the
        same key. Available to any authenticated user.
      tags: [push]
      responses:
        "200":
          description: The hub's VAPID public key.
          content:
            application/json:
              schema:
                type: object
                properties:
                  public_key: { type: string }
        "500":
          description: Failed to generate or persist a VAPID key pair.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/custom/push/test:
    post:
      operationId: sendTestPush
      summary: Send a test Web Push notification to the caller's own devices
      description: >
        Sends a test notification to every "push_subscriptions" row owned
        by the authenticated caller (never another user's), returning a
        per-endpoint success/failure result so the UI can show exactly
        which of the caller's devices received it. A subscription that the
        push service reports as gone (HTTP 404/410) is deleted as a side
        effect of this call. Recorded in the audit log as "push.test".
      tags: [push]
      responses:
        "200":
          description: Per-endpoint delivery results (possibly an empty list).
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        endpoint: { type: string }
                        success: { type: boolean }
                        error: { type: string }
        "500":
          description: Failed to load the caller's push subscriptions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A PocketBase auth token (superuser or "users" record) passed in the
        Authorization header. PocketBase accepts it either as a bare token
        or with a "Bearer " prefix.
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated, but the account's role is too low.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...